+++ src/couchdb/rolling-database.ts
@@ -408,11 +408,12 @@ export function makeRollingDatabase<T>(
     connection: ServerScope,
     design: string,
     view: string,
-    opts: RollingViewParams
+    opts: RollingViewParams & { chunkSize?: number }
   ): AsyncIterableIterator<CouchDoc<T>> {
     const {
       afterDate,
       partition,
+      chunkSize,
       // Native CouchDB options:
       ...rest
     } = opts
@@ -426,7 +427,7 @@ export function makeRollingDatabase<T>(
           : db.partitionedView(partition, design, view, allParams))
         return rows
       },
-      { afterDate }
+      { chunkSize, afterDate }
     )
   }
 
@@ -436,7 +437,9 @@ export function makeRollingDatabase<T>(
       if (database.startDate.valueOf() > date.valueOf()) continue
       return database.name
     }
-    throw new Error(`No rolling database exists for ${date.toISOString()}`)
+    throw new Error(
+      `No rolling database exists for ${name} at ${date.toISOString()}`
+    )
   }
 
   async function insert(
@@ -548,7 +551,7 @@ export function makeRollingDatabase<T>(
         }
         const { exists } = clusterHasDatabase(
           replicatorSetup?.doc,
-          pool.defaultName,
+          watchCluster,
           setup
         )
         if (exists) {