+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- added: Accept a chunk size for rolling database streaming queries.
+- fixed: Use the `watchCluster` to filter rolling databases, not the default pool cluster.
+
 ## 0.2.22 (2025-07-03)
 
 - fixed: Correctly ignore archived rolling databases.